added amber_reader_energy_cal_min.cpp.  this reads in cordenates and parameters and calculates the amber energy.
This reproduces the amber energies well.

We will add in a minimizer and my new squish force feild.

We will also add parameters to control if we will use the origenal force feild or the squishy force feild.

this is meant to be used for an aid to model building, not in preforming simulations.